# Cloud/contracts/framebase-api-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Framebase Development API",
    "version": "1.2.0",
    "description": "Private development cloud contract for Framebase Phase 4 organization and recovery."
  },
  "paths": {
...
    "/v1/auth/enroll": {
      "post": {
        "summary": "Development device enrollment",
        "parameters": [
          {
            "name": "X-Enrollment-Secret",
            "in": "header",
            "required": true
          }
        ]
      }
    },
...
    "/v1/auth/revoke": {
      "post": {
        "summary": "Revoke a temporary development device",
        "parameters": [
          {
            "name": "X-Enrollment-Secret",
            "in": "header",
            "required": true
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}